Publish first rudimentary implementation
I have started by specifying the API in the Swagger format. I have used
this specification to generate a server skeleton, based on the Spring
Boot framework. I have then made a number of adjustments in the
generated code (pom.xml and Java files).

I have also prepared an initial setup for Docker. There is a script to
generate the openaffect/java-server image, which encapsulates the
server. The script re-builds the .jar file, with the help of maven. I
have also defined a docker topology, by creating a Docker compose file.
In this file, I have 2 services: the Java server and a MongoDB server.

Finally, I have created a PostMan collection with 2 examples of API
calls. The first call shows how to report a measurement (a 'subject' has
feld an 'emotion' because of a certain 'trigger' at a certain time). The
second call shows how to get the list of all measurements previously
sent to the server.