Fix mistake in API spec and content-type issue
In the Swagger spec (.yaml), it is possible to define the supported
media types (e.g. application/json) at the API level, and then to
override them for every endpoint method.

The problem is that if we define 'consumes' at the API level, then when
the code is generated, the server expects to receive a 'Content-type'
header for every endpoint method. For GET methods, this does not make
sense.

The .yaml file has been modified because of this. The Postman example
has also been adapted.